requirements.txt
uvicorn
fastapi
sqlalchemy
pymysql
passlib
bcrypt
python-jose
python-multipart
